Add new cursor controls.
The old `termpaint_terminal_set_cursor` is not deprecated for use in a
running application. This call set the cursor position immediately. It
did not remember the set position, so after the next flush the cursor
was again in the lower right corner.

The newly added functions termpaint_terminal_set_cursor_position,
termpaint_terminal_set_cursor_visible and termpaint_terminal_set_cursor_style
all take effect on the next flush of the terminal and keep their values
until overriden by a new call.