termpaint: Add termpaint_terminal_pause and termpaint_terminal_unpause.
These allow to temporary halt termpaint mode. This can for example be
useful when invoking external programs that are expected to interact with
the terminal.

Also introduces termpaint_str as a simple string abstraction that
keeps track of already allocated memory to avoid reallocating the snippets
used in unpause for every change.