termpaintx_ttyrescue: Don't use PATH to find ttyrescue.
ttyrescue is an internal helper that actually does not belong into a
directory in the PATH. Instead search a configurable list of
directories.

This also reduces the risk of suid or otherwise privileged programs
executing a implant of some form, assuming the configured path only
includes system directories.

This does not yet implement an override for relocatable installations,
but that could be added in the future.