termpaintx_ttyrescue: Add support for fexec·ing ttyrescue internal binary from memory.
Using an fallback with fexec gives the ttyrescue process its own identity,
which helps against killall with the main executable's name. Also it should
use less resources because it does not keep old pages from the main
process around. This should offset the non shareable copy of the executable
in the memfd.