termpaint: Fix garbage collection when adding long text to cells.
termpaintp_hash_ensure has to be called on a consistant state of the cell
array. Changing cell::text_len to zero without also assigning a valid
hash_item will cause inline text to be interpreted as a pointer, invoking
undefined behavior.

Refactor all code to use a helper that implements the proper sequence.