termpaint: Add TERMPAINT_CAPABILITY_EXTENDED_CHARSET capability.
This capability tries to predict if a rich set of unicode codepoints
is available for display.
This capability is not currently used inside of termpaint, but can be used
by users of termpaint via termpaint_terminal_capable.

For now this is mostly to try to handle linux vc, that has a very limited
set of usable characters.