termpaint: Remove normalizing zeroed out cells to spaces with default colors.
The previous commit changed the numeric value of the default color to be
zero, so that part now is handled without patching things on first sight.

Additionally carry `text_len == 0 && text_overflow == nullptr` through as
distinct state and only translate that to space on terminal flush.

This allows removing a bunch of ugly code and opens up to use this as an
erased cell state.