termpaint_terminal_flush: Add support for terminal end of line heuristics.
Output runs of cleared cells at line end by using CSI K to clear the line
to trigger end of line heuristics in terminals.

This allows copying from the terminal to only include spaces that are not
from cleared cells. If the terminal uses compatible heuristics this allows
copying intended spaces at the end of a line but ignoring unintended spaces
that are just padding.