termpaint_surface_set_softwrap_marker,termpaint_terminal_flush: Add support for terminal soft wrap heuristics.
This allows creating soft wrapped lines that copy as one line from
terminals with suitable soft wrap heuristics.

termpaint_surface_set_softwrap_marker must be called by client code on the
last cell of the text of each wrapped physical line. Additionally it must
be called on the first cell of a continuation line. This convention was
selected because it hopefully plays nice with different parts of an client
using clipping and composing output of different components. It should
avoid joining lines accidentally except when composition by the clients
results in wrapped and continuation lines next to each other that belong to
a different underlying string.

Also change termpaint_terminal_flush to render this into the terminal with
a fairly simple strategy. This is not yet fully incremental, but the amount
of additional traffic is limited by the number of wrapped lines.