termpaint_input: Add simple bracketed paste support.
For now keep the structure that input is processed strictly one sequence or
character at a time. This is not ideal for paste, but higher layers can
reassemble the paste based on the initial and final flags. And serving the
whole paste in one event isn't possible without buffering anyway.