termpaint_surface_peek_text: Fix crash when left most character is double wide and called with x=1.
In this case the new value for cell was not generated because the while
condition was no longer fulfilled.

The getcell call at the top of the loop is not needed in the first iteration
because there is a getcell call above with the same parameters, so moving
at the end of the loop is safe and ensures that `cell` is refreshed with
the new x value.