meson.build,tests: Make sure `meson test` works.
Fix the main test by removing the hard coded path relation and use an
environment variable to find the test data.

Remove the output test from meson test for now, because it needs an
external terminal test driver. This would need to be at least a option
of the build, but ideally this should work with different drivers without
reconfiguring the build directory.